Make JAVA:JRESOLVE-METHOD try harder to resolve a JAVA-OBJECT instance.
Unclear if this behavior comes from a bug in how we intialize
JAVA-OBJECT's intendedClass or is a "natural" result of some casting
assumption.  We assume that users of JRESOLVE-METHOD would prefer some
sort of callable result as opposed to NIL as the current JAVA
introspection APIs assumption seems to geared to guessing a reasonable
default.  Those wishing strict introspection semantics are advised to
use the java.lang.reflect package directly.