Full build results from x64-pc-solaris-5.11.oi148.
Build metrics output for every revision from abcl-0.13 onwards.

Does not currently parse with build-metrics.lisp as there are abnormal
build results--at least one from a build.xml that was killed from an
endless loop--contained in the raw output file.

x64-pc-solaris-5.11.oi148 means the ABCL build host was a x86_x64
64bit processor running OpenSolaris as packaged by OpenIndiana from
build 148, which was released in at the end of 2010.  The oi148
instance is running on commercial multicore Intel chips with grownup
I/O and plenty of RAM.