Fill in some missing docstrings from JAVA package.
Unfortunately, we have to use static initializers and retain the '###'
marker for documenting package variables as we don't seem to be able
to use the @DocString annotation as the JVM reflection APIs only seem
to require that we know the Java class that a field is part of in
order to look up runtime annotations.  If this understanding is
faulty, please let me know otherwise how to do it.

The :DOCUMENTATION option for the DEFGENERIC is apparently not
working.