Fix Uniform Naming Convention (aka "UNC" or "network") paths under Windows.
DIRECTORY now works again on UNC paths.

UNC paths may be either specified with either back slash (#\\) or
forward slash (#\/) doubled as the first character in a Pathname
namestring.

The patterns in

        //<server>/<share>/[directories-and-files]

are parsed as

    <server> is stored as HOST.

    <share> is stored as DEVICE.

    [directories-and-files] gets parsed as per the normal rules under
    Windows.

Mixing namestrings with both backslash and slash characters can lead
to unpredictable results.  It is recommended not to use backslash
characters in namestrings if it can be avoided.  The pathname printed
representation is always normalized to using forward slash delimiters.